My App/requirements.txt
torch
torchvision
numpy
pillow
streamlit
streamlit-drawable-canvas
matplotlib

